refactor: code split - class Game
